requirements.txt
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
Flask==3.1.0
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
PyJWT==2.10.1
pytest==8.3.4
Werkzeug==3.1.3
